--- src/app/(root)/login/LoginClient.module.scss
@@ -1,144 +1,0 @@
-@import "@/styles/variables.scss";
-
-@font-face {
-  font-family: "YellowTail";
-  src: url("/fonts/Yellowtail-Regular.ttf");
-}
-
-.container {
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  overflow: hidden;
-  background-color: $input-color;
-}
-.card {
-  transition: 0.3s all ease-in-out !important;
-}
-.contents {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 100%;
-}
-
-.loginHeader {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 50px 0 2rem;
-}
-
-.loginHeader h1 {
-  font-family: "YellowTail";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 32px;
-  line-height: 24px;
-  text-align: center;
-  margin-bottom: 0;
-}
-
-.loginHeader p {
-  color: $grey-color;
-  font-size: 12px;
-  margin-bottom: 0;
-}
-
-.form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  gap: 1.5rem;
-  margin-top: 10px;
-}
-
-.inputField {
-  width: 320px;
-  height: 50px;
-  background: $input-color;
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-  border-radius: 15px;
-  border: none;
-  padding-left: 50px;
-  transition: 0.3s all ease-in-out !important;
-}
-
-.inputBox {
-  position: relative;
-}
-
-.icon {
-  position: absolute;
-  top: 16px;
-  left: 15px;
-}
-
-.eyeIcon {
-  position: absolute;
-  top: 15px;
-  right: 15px;
-}
-
-.loginButton {
-  margin-top: 40px;
-  font-weight: 700;
-  color: $primary-color !important;
-  transition: 0.3s all ease-in-out !important;
-
-  &:hover {
-    background-color: $primary-color !important;
-    color: white !important;
-  }
-}
-
-.loginMenu {
-  width: 370px;
-  font-weight: 700;
-  font-size: 13px;
-  padding-top: 10px;
-  display: flex;
-  gap: 10px;
-}
-
-.signup {
-  text-decoration: none;
-  color: $primary-color;
-  transition: 0.3s all ease-in-out !important;
-
-  &:hover {
-    color: $primary-color-hover !important;
-  }
-}
-
-.forgot {
-  text-decoration: none;
-  color: $grey-color;
-
-  &:hover {
-    color: $grey-color-hover !important;
-  }
-}
-
-@media (max-width: 580px) {
-  .card {
-    width: 400px !important;
-  }
-  .inputField {
-    width: 200px !important;
-  }
-  .loginButton {
-    width: 250px !important;
-  }
-  .loginMenu {
-    width: 250px;
-  }
-}
-@media (max-width: 480px) {
-  .card {
-    width: 280px !important;
-  }
-}
