+++ src/app/containers/login/LoginForm.module.scss
@@ -1,0 +1,51 @@
+@import "@/styles/variables.scss";
+.form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  gap: 1.5rem;
+  margin-top: 10px;
+}
+.inputBox {
+  position: relative;
+}
+.inputField {
+  width: 320px;
+  height: 50px;
+  background: $input-color;
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  border-radius: 15px;
+  border: none;
+  padding-left: 50px;
+  transition: 0.3s all ease-in-out !important;
+}
+.icon {
+  position: absolute;
+  top: 16px;
+  left: 15px;
+}
+.eyeIcon {
+  position: absolute;
+  top: 15px;
+  right: 15px;
+}
+.loginButton {
+  margin-top: 40px;
+  font-weight: 700;
+  color: $primary-color !important;
+  transition: 0.3s all ease-in-out !important;
+
+  &:hover {
+    background-color: $primary-color !important;
+    color: white !important;
+  }
+}
+@media (max-width: 580px) {
+  .inputField {
+    width: 200px !important;
+  }
+  .loginButton {
+    width: 250px !important;
+  }
+}
