+++ src/app/containers/login/index.module.scss
@@ -1,0 +1,31 @@
+@import "@/styles/variables.scss";
+
+.container {
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  overflow: hidden;
+  background-color: $input-color;
+}
+.card {
+  transition: 0.3s all ease-in-out !important;
+}
+.contents {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  height: 100%;
+}
+
+@media (max-width: 580px) {
+  .card {
+    width: 400px !important;
+  }
+}
+@media (max-width: 480px) {
+  .card {
+    width: 280px !important;
+  }
+}
